implemented the delete functionality

+++ src/main/java/com/example/employeemanagementsystem/service/EmployeeService.java
@@ -26,6 +26,8 @@ public class EmployeeService {
     }
 
     public void deleteEmployee(Long id) {
-        employeeRepository.deleteById(id);
+        if (employeeRepository.existsById(id)) {
+            employeeRepository.deleteById(id);
+        }
     }
 }
